requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
attrs==25.4.0
audioop-lts==0.2.2
brotli==1.2.0
cairocffi==1.7.1
CairoSVG==2.8.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
chyllonge==1.1.1
cssselect2==0.8.0
defusedxml==0.7.1
discord.py==2.6.4
frozenlist==1.8.0
idna==3.11
multidict==6.7.0
pillow==12.0.0
propcache==0.4.1
pycparser==2.23
python-dotenv==1.2.1
requests==2.32.5
setuptools==80.9.0
tinycss2==1.5.1
tzlocal==5.3.1
urllib3==2.6.2
webencodings==0.5.1
wheel==0.45.1
yarl==1.22.0
